refactor: :technologist: change Navigation Menu component to Smart and Dumb component concept
